refactor(request): extract request validation posting
