+++ src/handlers/request/index.ts
@@ -11,11 +11,6 @@ import {
   findOpenIssuePrs as findOpenIssuePRsRaw,
 } from './pr/snapshot.js';
 import { createRequestPr as createRequestPRRaw } from './pr/create.js';
-import {
-  buildMachineReadableMetadataBlock,
-  normalizeMachineReadableIssues,
-  type MachineReadableIssue,
-} from './domain/machine-readable.js';
 import { buildReviewHandoverBody as buildReviewHandoverBodyPure } from './domain/review-handover-rendering.js';
 import {
   ensureAutomatedApprovalReviewForCurrentHead as ensureAutomatedApprovalReviewForCurrentHeadApplication,
@@ -124,6 +119,10 @@ import {
   createRequestPrWithRecovery as createRequestPrWithRecoveryApplication,
   type RequestPrCreationRecoveryCallbacks,
 } from './application/request-pr-creation-recovery.js';
+import {
+  buildRegistryValidationAggregatePrCommentBody as buildRegistryValidationAggregatePrCommentBodyApplication,
+  type RequestValidationPostingCallbacks,
+} from './application/request-validation-posting.js';
 import {
   clearSequentialRegistryPrActive,
   getSequentialRegistryPrActive,
@@ -244,15 +243,7 @@ import {
   type ParentApprovalMeta,
 } from './domain/approval-markers.js';
 import { readIssueBodyForProcessing } from './domain/issue-body-processing.js';
-import {
-  buildRegistryValidationAggregateBody,
-  buildRegistryValidationCommentHeading,
-  extractFieldFromMsg,
-  filterMachineReadableSourcesForFile,
-  filterRegistryValidationEntries,
-  normalizeMsg,
-  type RegistryValidationMachineReadableSource,
-} from './domain/registry-validation-annotations.js';
+import { type RegistryValidationMachineReadableSource } from './domain/registry-validation-annotations.js';
 import { isApprovalComment, isAuthorUpdateComment, stripQuoteAndCode } from './domain/comment-commands.js';
 import { tryMergeIfGreen as tryMergeIfGreenRaw } from '../../lib/auto-merge.js';
 import { loadStaticConfig, DEFAULT_CONFIG, type NormalizedStaticConfig, type RegistryBotHooks } from '../../config.js';
@@ -438,8 +429,6 @@ type EffectiveConstants = {
   approverPoolUsernames: string[];
 };
 
-type SchemaFieldAliasLookup = Map<string, string>;
-
 type LogLevel = 'debug' | 'info' | 'warn' | 'error';
 type LoggerFn = (this: unknown, obj: unknown, msg?: string) => void;
 type LoggerLike = Partial<Record<LogLevel, LoggerFn>>;
@@ -452,160 +441,6 @@ const log = (context: { log?: LoggerLike } | undefined, level: LogLevel, obj: un
     fn.call(logger, obj, msg);
   }
 };
-
-const SCHEMA_FIELD_ALIAS_CACHE = new Map<string, Promise<SchemaFieldAliasLookup>>();
-
-function normalizeSchemaFieldAlias(value: unknown): string {
-  const raw = toStringTrim(value);
-  if (!raw) return '';
-
-  return raw
-    .replace(/([a-z0-9])([A-Z])/g, '$1 $2')
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '');
-}
-
-function addSchemaFieldAlias(lookup: SchemaFieldAliasLookup, aliasValue: unknown, propertyName: string): void {
-  const alias = normalizeSchemaFieldAlias(aliasValue);
-  if (!alias || lookup.has(alias)) return;
-
-  lookup.set(alias, propertyName);
-
-  if (alias.endsWith('s') && alias.length > 1) {
-    const singular = alias.slice(0, -1);
-    if (singular && !lookup.has(singular)) lookup.set(singular, propertyName);
-  } else {
-    const plural = `${alias}s`;
-    if (!lookup.has(plural)) lookup.set(plural, propertyName);
-  }
-}
-
-function collectSchemaFieldAliasesForProperty(
-  propertyName: string,
-  propertyDef: unknown,
-  lookup: SchemaFieldAliasLookup
-): void {
-  addSchemaFieldAlias(lookup, propertyName, propertyName);
-  if (!isPlainObject(propertyDef)) return;
-
-  addSchemaFieldAlias(lookup, propertyDef['title'], propertyName);
-  addSchemaFieldAlias(lookup, propertyDef['x-form-field'], propertyName);
-  collectSchemaFieldAliases(propertyDef, lookup);
-}
-
-function collectSchemaFieldAliasesFromProperties(props: Record<string, unknown>, lookup: SchemaFieldAliasLookup): void {
-  for (const [propertyName, propertyDef] of Object.entries(props)) {
-    collectSchemaFieldAliasesForProperty(propertyName, propertyDef, lookup);
-  }
-}
-
-function collectSchemaFieldAliasesFromArray(items: unknown[], lookup: SchemaFieldAliasLookup): void {
-  for (const item of items) {
-    collectSchemaFieldAliases(item, lookup);
-  }
-}
-
-function collectSchemaFieldAliases(schemaObj: unknown, lookup: SchemaFieldAliasLookup): void {
-  if (!isPlainObject(schemaObj)) return;
-
-  const props = isPlainObject(schemaObj['properties']) ? schemaObj['properties'] : null;
-  if (props) collectSchemaFieldAliasesFromProperties(props, lookup);
-
-  for (const key of ['allOf', 'anyOf', 'oneOf'] as const) {
-    const items = schemaObj[key];
-    if (!Array.isArray(items)) continue;
-
-    collectSchemaFieldAliasesFromArray(items, lookup);
-  }
-
-  const defs = isPlainObject(schemaObj['$defs']) ? schemaObj['$defs'] : null;
-  if (defs) {
-    for (const value of Object.values(defs)) {
-      collectSchemaFieldAliases(value, lookup);
-    }
-  }
-}
-
-async function loadSchemaFieldAliasLookup(
-  context: BotContext<RequestEvents>,
-  repoInfo: RepoInfo,
-  schemaPath: string
-): Promise<SchemaFieldAliasLookup> {
-  const rawPath = toStringTrim(schemaPath);
-  if (!rawPath) return new Map<string, string>();
-
-  const cleaned = rawPath.replace(/^\.?\//, '');
-  const candidates = rawPath.startsWith('/')
-    ? [rawPath.replace(/^\/+/, '')]
-    : [cleaned.startsWith('.github/') ? cleaned : `.github/registry-bot/${cleaned}`, cleaned];
-
-  const cacheKey = `${repoInfo.owner}/${repoInfo.repo}:${JSON.stringify(candidates)}`;
-  const cached = SCHEMA_FIELD_ALIAS_CACHE.get(cacheKey);
-  if (cached) return await cached;
-
-  const pending = (async (): Promise<SchemaFieldAliasLookup> => {
-    for (const candidate of candidates) {
-      const raw = await readRepoFileText(context, repoInfo, candidate);
-      if (!raw) continue;
-
-      try {
-        const parsed = JSON.parse(raw) as unknown;
-        const lookup = new Map<string, string>();
-        collectSchemaFieldAliases(parsed, lookup);
-        return lookup;
-      } catch {
-        continue;
-      }
-    }
-
-    return new Map<string, string>();
-  })();
-
-  SCHEMA_FIELD_ALIAS_CACHE.set(cacheKey, pending);
-  return await pending;
-}
-
-async function resolveMachineReadableRegistryField(
-  context: BotContext<RequestEvents>,
-  repoInfo: RepoInfo,
-  fieldHint: string,
-  schemaPath?: string
-): Promise<string> {
-  const fallback = toStringTrim(fieldHint) || 'details';
-  const normalizedSchemaPath = toStringTrim(schemaPath);
-
-  if (!normalizedSchemaPath || fallback === 'details') return fallback;
-
-  const lookup = await loadSchemaFieldAliasLookup(context, repoInfo, normalizedSchemaPath);
-  if (!lookup.size) return fallback;
-
-  return lookup.get(normalizeSchemaFieldAlias(fallback)) || fallback;
-}
-
-async function buildRegistryValidationMachineReadableIssues(
-  context: BotContext<RequestEvents>,
-  repoInfo: RepoInfo,
-  items: RegistryValidationMachineReadableSource[]
-): Promise<MachineReadableIssue[]> {
-  const out: MachineReadableIssue[] = [];
-
-  for (const item of items || []) {
-    const message = normalizeMsg(item.message);
-    if (!message) continue;
-
-    const fieldHint = extractFieldFromMsg(item.message) || 'details';
-    const field = await resolveMachineReadableRegistryField(context, repoInfo, fieldHint, item.schemaPath);
-    const normalizedFilePath = toStringTrim(item.filePath);
-
-    out.push({
-      field,
-      message,
-      ...(normalizedFilePath ? { filePath: normalizedFilePath } : {}),
-    });
-  }
-
-  return normalizeMachineReadableIssues(out);
-}
 
 function isPlainObject(value: unknown): value is Record<string, unknown> {
   return typeof value === 'object' && value !== null && !Array.isArray(value);
@@ -937,54 +772,28 @@ const normalizeKey = (s: unknown): string => {
   return base.replaceAll(/[^\w]+/g, '-').replaceAll(/(?:^-+|-+$)/g, '');
 };
 
-async function buildRegistryValidationPrCommentBody(
-  context: BotContext<RequestEvents>,
-  repoInfo: RepoInfo,
-  filePath: string,
-  messages: string[],
-  machineReadableSources: RegistryValidationMachineReadableSource[]
-): Promise<string> {
-  const lines: string[] = [
-    '## Detected issues',
-    '',
-    ...buildRegistryValidationCommentHeading(filePath, messages, '###'),
-  ];
-
-  const body = lines.join('\n').trimEnd();
-  const machineReadable = await buildRegistryValidationMachineReadableIssues(context, repoInfo, machineReadableSources);
-
-  return `${body}
-
-${buildMachineReadableMetadataBlock(machineReadable)}`;
-}
-
 async function buildRegistryValidationAggregatePrCommentBody(
   context: BotContext<RequestEvents>,
   repoInfo: RepoInfo,
   byFile: Map<string, string[]>,
   machineReadableSources: RegistryValidationMachineReadableSource[]
 ): Promise<string> {
-  const entries = filterRegistryValidationEntries(byFile);
+  return await buildRegistryValidationAggregatePrCommentBodyApplication(
+    context,
+    repoInfo,
+    byFile,
+    machineReadableSources,
+    buildRequestValidationPostingCallbacks()
+  );
+}
 
-  if (!entries.length) return '';
-  if (entries.length === 1) {
-    const [filePath, messages] = entries[0];
-    return await buildRegistryValidationPrCommentBody(
-      context,
-      repoInfo,
-      filePath,
-      messages,
-      filterMachineReadableSourcesForFile(machineReadableSources, filePath)
-    );
-  }
-
-  const body = buildRegistryValidationAggregateBody(byFile);
-
-  const machineReadable = await buildRegistryValidationMachineReadableIssues(context, repoInfo, machineReadableSources);
-
-  return `${body}
-
-${buildMachineReadableMetadataBlock(machineReadable)}`;
+function buildRequestValidationPostingCallbacks(): RequestValidationPostingCallbacks<
+  BotContext<RequestEvents>,
+  RepoInfo
+> {
+  return {
+    readRepoFileText,
+  };
 }
 
 const labelName = (l: unknown): string => {
